refactor/디렉토리구조-#88 : 컨벤션에 맞지 않은 디렉토리 구조를 정리하고, 관련 로직들을 관심사에 따라 분리한다
- cheer, protest API 경로 단순화 (`index.ts` → 단일 파일로)
- verification API 디렉토리 정리 및 파일 분리
- 기존 VerifyLocation.tsx 파일 삭제 후, API 호출과 로직 가공을 분리
- protest 관련 타입은 types/protest.ts로 이동
