+++ src/constants/protest.ts
@@ -1,0 +1,1 @@
+// protest constant
